requirements.txt
google-generativeai==0.7.2
streamlit==1.38.0
streamlit-option-menu==0.3.13
pillow==10.2.0